requirements.txt
pinecone-client
voyageai
anthropic
boto3